ESC-925 PR feedback updates
- move styling to styles.css
- fix typo in test name
- remove superfluous toLong() in Message

+++ koap/src/commonMain/kotlin/Message.kt
@@ -290,7 +290,7 @@ sealed class Message {
             )
 
             init {
-                require(value.toLong() in OBSERVE_RANGE) {
+                require(value in OBSERVE_RANGE) {
                     "Observe value of $value is outside allowable range of $OBSERVE_RANGE"
                 }
             }
